parsing

+++ matrix_create.h
@@ -6,7 +6,7 @@
 /*   By: trobicho <trobicho@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/11 20:11:28 by trobicho          #+#    #+#             */
-/*   Updated: 2019/04/15 17:33:15 by trobicho         ###   ########.fr       */
+/*   Updated: 2019/04/19 02:03:26 by trobicho         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,5 +17,5 @@
 #include "qlist.h"
 
 void	ft_relink_secondary(t_qlist *clst, int nb_primary);
-t_qlist	*ft_alloc_matrix(char **m, size_t w, size_t h, int nb_second);
+t_qlist	*ft_alloc_matrix(char **m, int w, int h, int nb_second);
 #endif
